deleted unused refactor warnings

+++ QZXing/QZXing.cpp
@@ -260,7 +260,7 @@ QRectF getTagRect(const ArrayRef<Ref<ResultPoint> > &resultPoints, const Ref<Bit
 
         qreal xMin = resultPoints[0]->getX();
         qreal xMax = xMin;
-        for (unsigned int i = 1; i < resultPoints->size(); ++i) {
+		for (int i = 1; i < resultPoints->size(); ++i) {
             qreal x = resultPoints[i]->getX();
             if (x < xMin)
                 xMin = x;
@@ -270,7 +270,7 @@ QRectF getTagRect(const ArrayRef<Ref<ResultPoint> > &resultPoints, const Ref<Bit
 
         qreal yMin = resultRectPoints[0]->getY();
         qreal yMax = yMin;
-        for (unsigned int i = 1; i < resultRectPoints.size(); ++i) {
+		for (size_t i = 1; i < resultRectPoints.size(); ++i) {
             qreal y = resultRectPoints[i]->getY();
             if (y < yMin)
                 yMin = y;
@@ -287,7 +287,7 @@ QRectF getTagRect(const ArrayRef<Ref<ResultPoint> > &resultPoints, const Ref<Bit
         qreal xMax = xMin;
         qreal yMin = resultPoints[0]->getY();
         qreal yMax = yMin;
-        for (unsigned int i = 1; i < resultPoints->size(); ++i) {
+		for (int i = 1; i < resultPoints->size(); ++i) {
             qreal x = resultPoints[i]->getX();
             qreal y = resultPoints[i]->getY();
             if (x < xMin)
@@ -485,8 +485,8 @@ QImage QZXing::encodeData(const QString& data)
         Ref<qrcode::ByteMatrix> bytesRef = barcode->getMatrix();
         const std::vector< std::vector <zxing::byte> >& bytes = bytesRef->getArray();
         image = QImage(bytesRef->getWidth(), bytesRef->getHeight(), QImage::Format_ARGB32);
-        for(int i=0; i<bytesRef->getWidth(); i++)
-            for(int j=0; j<bytesRef->getHeight(); j++)
+		for(size_t i=0; i<bytesRef->getWidth(); i++)
+			for(size_t j=0; j<bytesRef->getHeight(); j++)
                 image.setPixel(i, j, bytes[i][j] ?
                                    qRgb(0,0,0) :
                                    qRgb(255,255,255));
